Migrate TestHeroRecommendation.js
